+++ Changelog.md
@@ -4,6 +4,7 @@
 - New: Laser and inventory scanner fields can now be waterlogged, which means the Laser and Inventory Scanner now properly work underwater without air pockets
 - New: Randomize signature button for the Keycard Reader
 - Change: Several technical blocks' sounds have been adjusted to better match how they look
+- Change: Improved visuals when holding a taser
 - Fix: Shields do not take damage when blocking a taser that deals high enough damage
 - Fix: The death message of a player killed by a taser does not contain the player firing the taser
 - Fix: Crashes when rendering some modded block entities in a Projector, or using them as a disguise